feat: add profit and stock charts, enhance inventory context and filtering
- Introduced `getProfits` API to fetch profit data for the week.
- Added `HealthChart`, `ProfitChart`, and `StockChart` components for visualizing product health, profits, and stock distribution.
- Updated `CreateProductBtn` and `CreateProductModal` to utilize new inventory context for modal management.
- Implemented `FilterBtn` for filtering products by category and stock status.
- Enhanced `ProductStatsBanner` to use inventory context for displaying product statistics.
- Refactored `ProductTable` to integrate inventory context and improve product management actions.
- Created `InventoryProvider` to manage product selection, filtering, and modal states.
- Added `useFilter` hook for filtering products based on selected categories and stock status.
- Updated product update logic to handle stock adjustments with reasons.
- Modified CSS for improved theming and added utility for hiding scrollbars.
- Adjusted server-side `UpdateProductRequest` to use `BigDecimal` for stock values.

+++ server/src/main/java/com/express/inventory/dto/products/request/UpdateProductRequest.java
@@ -1,5 +1,7 @@
 package com.express.inventory.dto.products.request;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -9,7 +11,7 @@ public class UpdateProductRequest {
     private Integer id;
     private String name;
     private String productType;
-    private int stock;
+    private BigDecimal stock;
     private Double stockThreshold;
     private Double price;
